Fix waving text while autoscrolling
Package name refactoring
Increase autoscroll FPS

+++ app/src/main/java/igrek/songbook/domain/lyrics/LyricsLine.java
@@ -22,7 +22,7 @@ public class LyricsLine {
 		return fragments;
 	}
 	
-	public float getY() {
+	public int getY() {
 		return y;
 	}
